fixed weird compile warnings and errors
things I learned:
1. if you include string after iostream, you can't std::cout a std::string
2. lua_Number is double, not float

+++ LuaGame/LuaGame/Utility.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include <iostream>
+#include <string>
 #include <vector>
+#include <iostream>
 
 extern "C"
 {
@@ -16,7 +17,7 @@ namespace Utility
 	std::vector<std::string> SplitVariableName(lua_State* L, std::string varName);
 	void GetVariableOnStack(lua_State* L, std::string varName);
 	
-	float GetNumber(lua_State* L, std::string varName);
+	double GetNumber(lua_State* L, std::string varName);
 	std::string GetString(lua_State* L, std::string varName);
 
 	void LoadScript(lua_State* L, std::string scriptPath);
